+++ apps/researcher/src/lib/provenance-events-instance.ts
@@ -1,0 +1,8 @@
+import {ProvenanceEvents} from '@colonial-collections/api';
+import {env} from 'node:process';
+
+const provenanceEvents = new ProvenanceEvents({
+  sparqlEndpointUrl: env.SPARQL_ENDPOINT_URL as string,
+});
+
+export default provenanceEvents;
