+++ packages/api/src/index.ts
@@ -3,3 +3,4 @@ export * from './datasets';
 export * from './definitions';
 export * from './objects';
 export * from './organizations';
+export * from './provenance-events';
